+++ include/util/coordinate.hpp
@@ -68,19 +68,29 @@ using FloatLongitude = Alias<double, tag::longitude>;
 // range checks on these (toFixed/toFloat, etc)
 using UnsafeFloatLatitude = Alias<double, tag::unsafelatitude>;
 using UnsafeFloatLongitude = Alias<double, tag::unsafelongitude>;
-static_assert(std::is_standard_layout<FixedLatitude>() && std::is_trivial<FixedLatitude>(),
+static_assert(std::is_standard_layout<FixedLatitude>() &&
+                  std::is_trivially_default_constructible<FixedLatitude>() &&
+                  std::is_trivially_copyable<FixedLatitude>(),
               "FixedLatitude is not a valid alias");
-static_assert(std::is_standard_layout<FixedLongitude>() && std::is_trivial<FixedLongitude>(),
+static_assert(std::is_standard_layout<FixedLongitude>() &&
+                  std::is_trivially_default_constructible<FixedLongitude>() &&
+                  std::is_trivially_copyable<FixedLongitude>(),
               "FixedLongitude is not a valid alias");
-static_assert(std::is_standard_layout<FloatLatitude>() && std::is_trivial<FloatLatitude>(),
+static_assert(std::is_standard_layout<FloatLatitude>() &&
+                  std::is_trivially_default_constructible<FloatLatitude>() &&
+                  std::is_trivially_copyable<FloatLatitude>(),
               "FloatLatitude is not a valid alias");
-static_assert(std::is_standard_layout<FloatLongitude>() && std::is_trivial<FloatLongitude>(),
+static_assert(std::is_standard_layout<FloatLongitude>() &&
+                  std::is_trivially_default_constructible<FloatLongitude>() &&
+                  std::is_trivially_copyable<FloatLongitude>(),
               "FloatLongitude is not a valid alias");
 static_assert(std::is_standard_layout<UnsafeFloatLatitude>() &&
-                  std::is_trivial<UnsafeFloatLatitude>(),
+                  std::is_trivially_default_constructible<UnsafeFloatLatitude>() &&
+                  std::is_trivially_copyable<UnsafeFloatLatitude>(),
               "UnsafeFloatLatitude is not a valid alias");
 static_assert(std::is_standard_layout<UnsafeFloatLongitude>() &&
-                  std::is_trivial<UnsafeFloatLongitude>(),
+                  std::is_trivially_default_constructible<UnsafeFloatLongitude>() &&
+                  std::is_trivially_copyable<UnsafeFloatLongitude>(),
               "UnsafeFloatLongitude is not a valid alias");
 
 /**